requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
joblib==1.2.0
Markdown==3.4.1
nltk==3.7
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
sqlparse==0.4.3
textblob==0.17.1
tqdm==4.64.1
urllib3==1.26.13
vaderSentiment==3.3.2
whitenoise==6.2.0
